refactor: highlight 유틸 함수들 파일별로 분리
