+++ frontend/src/hooks/highlight/useCheckHighlight.tsx
@@ -1,12 +1,12 @@
 import { useState } from 'react';
 
 import { HIGHLIGHT_SPAN_CLASS_NAME, SENTENCE_BASIC_CLASS_NAME } from '@/constants';
-import { SelectionInfo } from '@/utils';
+import { EditorSelectionInfo } from '@/utils';
 
 const useCheckHighlight = () => {
   const [isAddingHighlight, setIsAddingHighlight] = useState(false);
 
-  const checkHighlight = (info: SelectionInfo) => {
+  const checkHighlight = (info: EditorSelectionInfo) => {
     const selectedAllSpanList = getAllSpanInSelection(info.selection);
     const isNoneHighlight = selectedAllSpanList.some((span) => !span.classList.contains(HIGHLIGHT_SPAN_CLASS_NAME));
